pyproject.toml
[project]
name = "movie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
"requests>=2.31.0",
    "fake-useragent>=1.4.0",
    "beautifulsoup4>=4.13.5",
    "sqlalchemy>=2.0.0",
    "requests-html>=0.10.0",
    "toml>=0.10.2",
    "singleton-decorator>=1.0.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0.0",
    "colorlog>=6.8.0",
]
